refactor: consilidate into v3.go
